Perbaikan Controller yang menggunkana Auth

+++ controllers/MhsAuthController.php
@@ -2,12 +2,11 @@
 
 namespace App\Controllers;
 
-include __DIR__.'/../config/connect.php';
+include __DIR__.'/../config/App.php';
+
 use Config\Connect as Connect;
-
-include __DIR__.'/../handler/Response.php';
+use Handler\CheckAuth as CheckAuth;
 use Handler\Response as Response;
-
 
 
 class MhsAuthController {
@@ -20,26 +19,35 @@ class MhsAuthController {
 
 	      $mysqli = Connect::mariadb();
 
-	      $qry = "SELECT * FROM mhs";
+	  	  if(CheckAuth::auth($request['token'])){
+
+		      $qry = "SELECT * FROM mhs";
+		      
+		      $exec = mysqli_query($mysqli, $qry);
+		     
+		      while($d = mysqli_fetch_array($exec)){
+
+		        $data[] = [
+			                   "id"          => $d['id'],
+			                   "nama"        => $d['nama'],
+			                   "jurusan"     => $d['jurusan'],
+			                   "fakultas"    => $d['fakultas'],
+			                   "nama_kampus" => $d['nama_kampus']
+			                 ]; 
+
+			    $code = "200";
+			    $msg  = "OK";
+
+		      }
+
+	       }else{
+		
+				$code = "401";
+				$msg  = "Token Invalid";
+	   	 
+	   	   }
 	      
-	      $exec = mysqli_query($mysqli, $qry);
-	     
-	      while($d = mysqli_fetch_array($exec)){
-
-	        $data[] = [
-		                   "id"          => $d['id'],
-		                   "nama"        => $d['nama'],
-		                   "jurusan"     => $d['jurusan'],
-		                   "fakultas"    => $d['fakultas'],
-		                   "nama_kampus" => $d['nama_kampus']
-		                 ]; 
-
-		    $code = "200";
-
-	      }
-
-	      
-		  return Response::res_auth($code, $msg=[], $data, $request);
+		  return Response::res($code, $msg=[], $data, $request);
     
     } catch(Exception $e){
       echo "error: {$e->getMessage()}";
@@ -55,25 +63,36 @@ class MhsAuthController {
 
 	      $mysqli = Connect::mariadb();
 
-	      $qry = "SELECT * FROM mhs WHERE nim = '$request[nim]' LIMIT 1";
-	      
-	      $exec = mysqli_query($mysqli, $qry);
-	     
-	      while($d = mysqli_fetch_array($exec)){
+	  	  if(CheckAuth::auth($request['token'])){
+			 
+		      $qry = "SELECT * FROM mhs WHERE nim = '$request[nim]' LIMIT 1";
+		      
+		      $exec = mysqli_query($mysqli, $qry);
+		     
+		      while($d = mysqli_fetch_array($exec)){
 
-	        $data[] = [
-		                   "id"          => $d['id'],
-		                   "nama"        => $d['nama'],
-		                   "jurusan"     => $d['jurusan'],
-		                   "fakultas"    => $d['fakultas'],
-		                   "nama_kampus" => $d['nama_kampus']
-		                 ]; 
+		        $data[] = [
+			                   "id"          => $d['id'],
+			                   "nama"        => $d['nama'],
+			                   "jurusan"     => $d['jurusan'],
+			                   "fakultas"    => $d['fakultas'],
+			                   "nama_kampus" => $d['nama_kampus']
+			                 ]; 
 
-		    $code = "200";
+			    $code = "200";
+			    $msg  = "OK";
 
-	      }
+		      }
+	
+	       }else{
+		
+				$code = "401";
+				$msg  = "Token Invalid";
+	   	 
+	   	   }
 
-		  return Response::res_auth($code, $msg=[], $data, $request=[]);
+
+		  return Response::res($code, $msg, $data, $request);
 
     } catch(Exception $e){
       echo "error: {$e->getMessage()}";
@@ -94,24 +113,33 @@ class MhsAuthController {
 	  	 $fakultas    = $request['fakultas'];
 	  	 $nama_kampus = $request['nama_kampus'];
 
-		 $result = mysqli_query($mysqli, "INSERT INTO mhs 
-		 								   (id,nama,nim,jurusan,fakultas,nama_kampus) 
-		 								  VALUES 
-		 								   (null,'$nama','$nim','$jurusan','$fakultas','$nama_kampus')
-		 						"); 
+	  	 if(CheckAuth::auth($request['token'])){
+			 
+			 $result = mysqli_query($mysqli, "INSERT INTO mhs 
+			 								   (id,nama,nim,jurusan,fakultas,nama_kampus) 
+			 								  VALUES 
+			 								   (null,'$nama','$nim','$jurusan','$fakultas','$nama_kampus')
+			 						"); 
 
-		 if($result){ 
-			$code = "200";
-			$msg  = "Data Mahasiswa Berhasil Di Input";
-		 }else{
-			$code = "500";
-			$msg  = "Data Mahasiswa Gagal Di Input";
-	     }
+			 if($result){ 
+				$code = "200";
+				$msg  = "Data Mahasiswa Berhasil Di Input";
+			 }else{
+				$code = "500";
+				$msg  = "Data Mahasiswa Gagal Di Input";
+		     }
 	   
-		 return Response::res_auth($code, $msg, $data=[], $request);
+	   	 }else{
+		
+				$code = "401";
+				$msg  = "Token Invalid";
+	   	 
+	   	 }
+
+		 return Response::res($code, $msg, $data=[], $request);
     
     } catch(Exception $e){
-      echo "error: {$e->getMessage()}";
+      echo $e;
     }  
 
 	     
@@ -132,24 +160,34 @@ class MhsAuthController {
 	  	 $fakultas    = $request['fakultas'];
 	  	 $nama_kampus = $request['nama_kampus'];
 
-		 $result = mysqli_query($mysqli, "UPDATE mhs 
-		 								   SET nama = '$nama',
-		 								       nim  = '$nim',
-		 								       jurusan  = '$jurusan',
-		 								       fakultas  = '$fakultas',
-		 								       nama_kampus  = '$nama_kampus'
-		 								   WHERE id = '$id'
-		 						"); 
 
-		 if($result){ 
-			$code = "200";
-			$msg  = "Data Mahasiswa Berhasil Di Perbaharui";
-		 }else{
-			$code = "500";
-			$msg  = "Data Mahasiswa Gagal Di Perbaharui";
-	     }
+	  	 if(CheckAuth::auth($request['token'])){
+			 
+			 $result = mysqli_query($mysqli, "UPDATE mhs 
+			 								   SET nama = '$nama',
+			 								       nim  = '$nim',
+			 								       jurusan  = '$jurusan',
+			 								       fakultas  = '$fakultas',
+			 								       nama_kampus  = '$nama_kampus'
+			 								   WHERE id = '$id'
+			 						"); 
+
+			 if($result){ 
+				$code = "200";
+				$msg  = "Data Mahasiswa Berhasil Di Perbaharui";
+			 }else{
+				$code = "500";
+				$msg  = "Data Mahasiswa Gagal Di Perbaharui";
+		     }
 	   
-		 return Response::res_auth($code, $msg, $data=[], $request);
+	   	 }else{
+		
+				$code = "401";
+				$msg  = "Token Invalid";
+	   	 
+	   	 }
+
+		 return Response::res($code, $msg, $data=[], $request);
     
     } catch(Exception $e){
       echo "error: {$e->getMessage()}";
@@ -171,19 +209,28 @@ class MhsAuthController {
 
 	  	 $id     = $request['id'];
 
-		 $result = mysqli_query($mysqli, "DELETE FROM mhs
-		 								   WHERE id = '$id'
-		 						"); 
+	  	 if(CheckAuth::auth($request['token'])){
 
-		 if($result){ 
-			$code = "200";
-			$msg  = "Data Mahasiswa Berhasil Di Hapus";
-		 }else{
-			$code = "500";
-			$msg  = "Data Mahasiswa Gagal Di Hapus";
-	     }
+			 $result = mysqli_query($mysqli, "DELETE FROM mhs
+			 								   WHERE id = '$id'
+			 						"); 
+
+			 if($result){ 
+				$code = "200";
+				$msg  = "Data Mahasiswa Berhasil Di Hapus";
+			 }else{
+				$code = "500";
+				$msg  = "Data Mahasiswa Gagal Di Hapus";
+		     }
 	   
-		 return Response::res_auth($code, $msg, $data=[], $request);
+	   	 }else{
+		
+				$code = "401";
+				$msg  = "Token Invalid";
+	   	 
+	   	 }
+
+		 return Response::res($code, $msg, $data=[], $request);
     
     } catch(Exception $e){
       echo "error: {$e->getMessage()}";
@@ -204,31 +251,40 @@ class MhsAuthController {
 
 	  	 $key    = $request['key'];
 
-		 $result = mysqli_query($mysqli, "SELECT * FROM mhs
-		 								   WHERE nama LIKE '%$key%'
-		 						"); 
+	  	 if(CheckAuth::auth($request['token'])){
+		  	 	
+			 $result = mysqli_query($mysqli, "SELECT * FROM mhs
+			 								   WHERE nama LIKE '%$key%'
+			 						"); 
 
-	      while($d = mysqli_fetch_array($result)){
+		      while($d = mysqli_fetch_array($result)){
 
-	        $data[] = [
-		                   "id"          => $d['id'],
-		                   "nama"        => $d['nama'],
-		                   "nim"         => $d['nim'],
-		                   "jurusan"     => $d['jurusan'],
-		                   "fakultas"    => $d['fakultas'],
-		                   "nama_kampus" => $d['nama_kampus']
-		                 ]; 
-	      }
+		        $data[] = [
+			                   "id"          => $d['id'],
+			                   "nama"        => $d['nama'],
+			                   "nim"         => $d['nim'],
+			                   "jurusan"     => $d['jurusan'],
+			                   "fakultas"    => $d['fakultas'],
+			                   "nama_kampus" => $d['nama_kampus']
+			                 ]; 
+		      }
 
-		 if(mysqli_num_rows($result) > 0) { 
-			$code = "200";
-			$msg  = "Terdeteksi ".mysqli_num_rows($result)." Mahasiswa Yang Anda Cari..";
-		 }else{
-			$code = "404";
-			$msg  = "Data Mahasiswa Tidak Ditemukan";
-	     }
+			 if(mysqli_num_rows($result) > 0) { 
+				$code = "200";
+				$msg  = "Terdeteksi ".mysqli_num_rows($result)." Mahasiswa Yang Anda Cari..";
+			 }else{
+				$code = "404";
+				$msg  = "Data Mahasiswa Tidak Ditemukan";
+		     }
 	   
-		 return Response::res_auth($code, $msg, $data, $request);
+	   	 }else{
+		
+				$code = "401";
+				$msg  = "Token Invalid";
+	   	 
+	   	 }
+
+		 return Response::res($code, $msg, $data, $request);
     
     } catch(Exception $e){
       echo "error: {$e->getMessage()}";
